Add DonationRequest DTO,InterFace , Service,Controller and ad Doner Git profile and Rquest

+++ BloodDonationPlatform.API/Services/DTOs/DonationRequest/CreateDonationRequestDTO.cs
@@ -4,6 +4,6 @@
     {
         public int HospitalId { get; set; }
         public int BloodTypeId { get; set; }
-        public decimal NumOfLiter { get; set; }
+        public int NumOfLiter { get; set; }
     }
 }
